requirements.txt
Django>=4.0
psycopg2-binary
django-apscheduler
djangorestframework
python-dateutil
prometheus_client
google-generativeai>=0.3.0
python-dotenv
django-cors-headers